requirements.txt
certifi==2017.4.17
chardet==3.0.4
dj-database-url==0.4.1
Django==1.11.1
django-filter==1.0.4
djangorestframework==3.6.3
gunicorn==19.6.0
idna==2.5
Markdown==2.6.8
psycopg2==2.6.2
PyJWT==1.5.0
PySocks==1.6.7
pytz==2017.2
requests==2.18.1
six==1.10.0
twilio==6.4.2
urllib3==1.21.1
whitenoise==3.2
